Move experence section into components folder
